♻️ Refactor: Moved AemConfigManager to adapters
The AemInstaller sociable tests required a lot of mocking for
AemConfigManager.  This is an indication that it was not is the correct
place (as Victor Renta would say "listen to what your tests are telling
you").

By making it an adapter, I can now test using WireMock instead of
Mockito, which should make the tests more succinct.  Also, the sociable
test should require less mocking which will make it less fragile to
changes.