✅ Created RestClientAemConfigManager test that uses Wiremock
Created this to compare with the unit test that uses Mockito to mock the
RestClient interface.

Still decoding whether this is better than the unit test that uses
Mockito to mock the RestClient.

Leaving both in for now  with the expectation that one of these two will
be deleted at some future date.