✨ Change CommonsIoTailerTailer to use virtual threads
Hopefully this will address an issue with the CommonsIoTailerTailerTests
sometimes failing under GitHub Actions but not on local builds.

My hypothesis is that the GitHub Actions runner only has one core (as
opposed to mant on local build machines).  This means that the
multi-threaded code in CommonsIoTailerTailerTests is sometimes being
starved because platform threads are blocking on IO which, in turn,
causes timeouts.  The switch to virtual threads should prevent this and
is an overall better approach anyway.